refactor(dashboard-nav): split flagged components under 150 lines (#196)
Behavior-preserving refactor of dashboard & navigation components to
satisfy max-lines-per-function (max 150):

- DashboardMain: move animation CSS to a module-scope styled-jsx css
  block, extract DashboardStatCards and RecentDemos, hoist the editor
  query builder to module scope.
- SidemenuDashboard: extract SidebarBrand and a data-driven SidebarNav.
- SignedHeader: extract HeaderTitleIcon, ThemeToggleButton, UserMenu and
  a useProfileImage hook (dedupes the profile-image fetch effects).
- AnalyticsMain: extract AnalyticsStatCards, ViewsOverTimePanel and
  TopDemosPanel.
